+++ iModelCore/ECDb/PublicAPI/ECDb/ECDb.h
@@ -124,11 +124,11 @@ struct ECSqlConfig {
     private:
         DisableSqlFunctions m_disabledFunctions;
         bool m_experimentalFeaturesEnabled;
-        bool m_validateInsertValues;
+        bool m_validateWriteValues;
         mutable std::unordered_map<OptimizationOptions, bool> m_optimisationOptionsMap;
 
     public:
-        ECSqlConfig(): m_experimentalFeaturesEnabled(false), m_validateInsertValues(false) {
+        ECSqlConfig(): m_experimentalFeaturesEnabled(false), m_validateWriteValues(false) {
             m_optimisationOptionsMap[OptimizationOptions::OptimizeJoinForClassIds] = true;
             m_optimisationOptionsMap[OptimizationOptions::OptimizeJoinForNestedSelectQuery] = true;
         }
@@ -140,8 +140,8 @@ struct ECSqlConfig {
         bool GetExperimentalFeaturesEnabled() const { return m_experimentalFeaturesEnabled; }
         void SetExperimentalFeaturesEnabled(bool v)  { m_experimentalFeaturesEnabled = v; }
 
-        bool IsInsertValueValidationEnabled() const { return m_validateInsertValues; }
-        void SetInsertValueValidation(const bool v) { m_validateInsertValues = v; }
+        bool IsWriteValueValidationEnabled() const { return m_validateWriteValues; }
+        void SetWriteValueValidation(const bool v) { m_validateWriteValues = v; }
 };
 
 //=======================================================================================